feat(process): add path parameter in `wrap` methods
Added support for an optional `paths` parameter in the `wrap` methods of both `WorkerService` and `WorkerPool` classes, enhancing module resolution flexibility. This change allows users to specify an array of paths that can be used when resolving the module ID, facilitating the use of dynamic imports and module wrapping in worker threads. The introduction of the `toArray` utility function ensures that the `paths` parameter can be seamlessly handled as either a string or an array of strings, improving the API's usability.

Additionally, removed deprecated constants file `createWorkerService.constants.ts` and integrated `getWorkerServicePath` function directly into `createWorkerService.ts` to dynamically resolve the worker service script path, streamlining the worker service initialization process.

This update promotes better modularity and flexibility in handling worker threads within the `process` package, reflecting an ongoing commitment to enhancing the developer experience and capabilities of the library.