chore(eslint-config): update ESLint Vue configuration rules
Updated the ESLint configuration for Vue to enhance code readability and maintainability. This includes disabling the `no-unused-vars` rule to avoid false positives with the Composition API, adjusting rules for consistent spacing and newlines in templates, and enforcing consistent usage of type imports. Additionally, rules for HTML comments formatting and spacing between HTML/Component tags were refined to improve the structure and readability of Vue components.