chore(eslint-config): update rules for Node.js imports and version compatibility
Update the ESLint configuration for Node.js projects to enable built-in rules up to version 20 of Node.js, allowing the use of newer ECMAScript features and APIs. Additionally, the `n/prefer-node-protocol` rule has been adjusted to enforce the use of the `node:` protocol for importing Node.js built-in modules with a version specification. The `n/no-missing-import` rule description has been updated for clarity.