feat(unocss-preset): renamed preset to match current project naming
Renamed the preset from `presetHsjm` to `presetUnshared` and updated the interface from `PresetHsjmOptions` to `PresetUnshared`. This change aligns the preset's name with the current project naming conventions, removing the leftover name from the old project. Additionally, adjusted the order of properties within the preset to logically group related configurations, enhancing the readability and maintainability of the code.