feat(vue): prefix component names with 'Base'
Renamed several component files and their export names to include the 'Base' prefix. This change applies to `Alerts`, `Button`, and `InputToggle` components which are now named `BaseAlerts`, `BaseButton`, and `BaseInputToggle` respectively. The renaming is part of a larger effort to standardize component names and make their purpose and hierarchy clearer within the library.