chore: update .git-blame-ignore-revs with recent commit hashes
Updated the `.git-blame-ignore-revs` file to include hashes of recent commits that involved significant code style changes. This will make the blame tool more useful by ignoring these commits, which are primarily style changes and do not affect the functionality of the code. This helps in tracking meaningful changes and simplifies understanding the history of functional changes.